requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
idna==3.3
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
urllib3==1.26.9
